[BREAKINGCHANGE] Rename ChartsThemeProvider to ChartsProvider
