+++ ui/components/src/StatChart/calculateFontSize.ts
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { useChartsTheme } from '../context/ChartsThemeProvider';
+import { useChartsTheme } from '../context/ChartsProvider';
 import { FontSizeOption } from '../FontSizeSelector';
 
 interface CalculateFontSize {
