+++ ui/components/src/index.ts
@@ -37,9 +37,7 @@ export * from './TimeChart';
 export * from './TimeSeriesTooltip';
 export * from './UnitSelector';
 export * from './YAxisLabel';
-export * from './context/ChartsThemeProvider';
-export * from './context/TimeZoneProvider';
-export * from './context/SnackbarProvider';
+export * from './context';
 export * from './utils';
 export * from './model';
 export * from './test-utils';
