+++ ui/panels-plugin/src/plugins/time-series-chart/TimeSeriesChartPanel.tsx
@@ -89,7 +89,7 @@ export function TimeSeriesChartPanel(props: TimeSeriesChartProps) {
 
   const chartRef = useRef<ChartInstance>(null);
 
-  // ECharts theme comes from ChartsThemeProvider, more info: https://echarts.apache.org/en/option.html#color
+  // ECharts theme comes from ChartsProvider, more info: https://echarts.apache.org/en/option.html#color
   // Colors are manually applied since our legend and tooltip are built custom with React.
   const categoricalPalette = chartsTheme.echartsTheme.color;
 
@@ -193,7 +193,7 @@ export function TimeSeriesChartPanel(props: TimeSeriesChartProps) {
 
         // Color is used for line, tooltip, and legend
         const seriesColor = getSeriesColor({
-          // ECharts type for color is not always an array but it is always an array in ChartsThemeProvider
+          // ECharts type for color is not always an array but it is always an array in ChartsProvider
           categoricalPalette: categoricalPalette as string[],
           visual,
           muiPrimaryColor: muiTheme.palette.primary.main,
